+++ src/platform/apple/methods/file/fs/file_metadata/permissions.rs
@@ -4,10 +4,10 @@ use crate::flags::PermissionFlags;
 
 impl FileMetadata {
     pub fn permissions(&self) -> Permissions {
-         return Permissions::from_raw(self.metadata.st_mode as usize)
+         return Permissions::from_raw(self.metadata.st_mode)
     }
 
     pub fn permission_flags(&self) -> PermissionFlags {
-        return PermissionFlags::from_bits_truncate(self.metadata.st_mode as usize & PermissionFlags::MaskType.bits())
+        return PermissionFlags::from_bits_truncate(self.metadata.st_mode as u32 & PermissionFlags::MaskType.bits())
     }
 }
